+++ src/components/Graph/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-11-16: Initial version.
+- 2025-11-23: Refactor the code to use constants. Let the edges use bezier
+  curves. Support dark mode.